tests/wpt/tests/css/css-transforms/parsing/scale-parsing-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Transform Module Level 2: parsing scale with valid values</title>
<link rel="help" href="https://drafts.csswg.org/css-transforms-2/#propdef-scale">
<meta name="assert" content="scale supports the full grammar 'none | <number>{1,3}'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("scale", "none");

test_valid_value("scale", "1");
test_valid_value("scale", "1%", "0.01");

test_valid_value("scale", "100");
test_valid_value("scale", "100%", "1");
test_valid_value("scale", "100 100", "100");
test_valid_value("scale", "100% 100%", "1");
test_valid_value("scale", "100 100 1", "100");
test_valid_value("scale", "100% 100% 1", "1");

test_valid_value("scale", "-100");
test_valid_value("scale", "-100%", "-1");
test_valid_value("scale", "-100 -100", "-100");
test_valid_value("scale", "-100% -100%", "-1");
test_valid_value("scale", "-100 -100 1", "-100");
test_valid_value("scale", "-100% -100% 1", "-1");

test_valid_value("scale", "100 200");
test_valid_value("scale", "100% 200%", "1 2");
test_valid_value("scale", "100 200 1", "100 200");
test_valid_value("scale", "100% 200% 1", "1 2");
test_valid_value("scale", "100 200 300");
test_valid_value("scale", "100 100 2", "100 100 2");
test_valid_value("scale", "100% 200% 300%", "1 2 3");

// calc() expressions preserve units
test_valid_value("scale", "calc(200%)");
test_valid_value("scale", "calc(2)");
test_valid_value("scale", "calc(200%) calc(3)");
test_valid_value("scale", "calc(2 * 1) 3 calc(4 * 100%)", "calc(2) 3 calc(400%)");
test_valid_value("scale", "calc(-100%)");

// calc() expressions that are resolvable at computed value time
test_valid_value("scale", "calc(200% * sign(1em - 1px))");
test_valid_value("scale", "calc(2 * sign(1em - 1px))");
test_valid_value("scale", "calc(200%) calc(3 * sign(1em - 1px))");
test_valid_value("scale", "calc(2 * sign(1em - 1px)) 3 calc(4 * 100%)", "calc(2 * sign(1em - 1px)) 3 calc(400%)");
test_valid_value("scale", "calc(-100% * sign(1em - 1px))");
</script>
</body>
</html>
