third_party/blink/web_tests/external/wpt/css/css-transforms/parsing/scale-parsing-valid-expected.txt

This is a testharness.js-based test.
PASS e.style['scale'] = "none" should set the property value
PASS e.style['scale'] = "1" should set the property value
PASS e.style['scale'] = "1%" should set the property value
PASS e.style['scale'] = "100" should set the property value
PASS e.style['scale'] = "100%" should set the property value
PASS e.style['scale'] = "100 100" should set the property value
PASS e.style['scale'] = "100% 100%" should set the property value
FAIL e.style['scale'] = "100 100 1" should set the property value assert_equals: serialization should be canonical expected "100" but got "100 100 1"
FAIL e.style['scale'] = "100% 100% 1" should set the property value assert_equals: serialization should be canonical expected "1" but got "1 1 1"
PASS e.style['scale'] = "-100" should set the property value
PASS e.style['scale'] = "-100%" should set the property value
PASS e.style['scale'] = "-100 -100" should set the property value
PASS e.style['scale'] = "-100% -100%" should set the property value
FAIL e.style['scale'] = "-100 -100 1" should set the property value assert_equals: serialization should be canonical expected "-100" but got "-100 -100 1"
FAIL e.style['scale'] = "-100% -100% 1" should set the property value assert_equals: serialization should be canonical expected "-1" but got "-1 -1 1"
PASS e.style['scale'] = "100 200" should set the property value
PASS e.style['scale'] = "100% 200%" should set the property value
FAIL e.style['scale'] = "100 200 1" should set the property value assert_equals: serialization should be canonical expected "100 200" but got "100 200 1"
FAIL e.style['scale'] = "100% 200% 1" should set the property value assert_equals: serialization should be canonical expected "1 2" but got "1 2 1"
PASS e.style['scale'] = "100 200 300" should set the property value
PASS e.style['scale'] = "100 100 2" should set the property value
PASS e.style['scale'] = "100% 200% 300%" should set the property value
Harness: the test ran to completion.
